feat(transform): Add bad-requext-exception transformer to respect SRP
